Move encounter and weapon tests into separate folders
